microservices/waste-items-service/requirements.txt
passlib
fastapi
pymongo
python-jose
uvicorn